% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUANT.R
\name{predict.squant}
\alias{predict.squant}
\title{SQUANT prediction}
\usage{
\method{predict}{squant}(object, data, ...)
}
\arguments{
\item{object}{The squant object, the signature of which will be applied
to the specified data. The output of \code{squant} function.}

\item{data}{The data frame for prediction.}

\item{...}{Ignored.}
}
\value{
A data frame with the predicted subgroup in the last column.
}
\description{
\code{predict} assigns subgroup for each individual in a new dataset.
}
\details{
This function assigns subgroup for each individual in a new dataset
based on the derived signature contained within the squant object.
}
